refactor: del og flytt human_delta til utils
